Summarize Solid World September in TMiS
In an effort to make This Month in Solid more informative for recipients who missed Solid World, I have added an extended summary of the Solid World September presentations, and the upcoming event in Flanders. And I added two items to "Solid Implementations" that were mentioned as part of the Solid World presentations.